add execution of modulefiles from a modulefile
Some modulefiles may execute another modulefile,
which may modify the PATH variable. This is
generally done with the dofile() or loadfile()
function, but they break out of the sandbox.
By redefining these functions so the file is
executed in the sandbox env, modulefiles can
be safely loaded from within another modulefile.