add hierarchical modulepath support
if MII_ENABLE_SPIDER is set to yes at build time, mii
will use the spider command provided by Lmod to find
every modulefile in the modulepath hierarchy. This
method is not very compatible with syncing, because
the spider command scans every modulefile and then
outputs everything in JSON format.

MII_ENABLE_LUA can also be used in conjunction with
MII_ENABLE_SPIDER, but will only have an effect when
doing a sync, because the spider command will be used
when doing a build.