add global index support using env var
Default behaviour remains unchanged. If MII_INDEX_FILE
is set, it will be used as the _mii_datafile. The
behaviour for _mii_datadir also remains unchanged,
so disabling and enabling mii can be done on a
per-user basis.

Also, disabled sync on sourcing shell integration if
MII_INDEX_FILE is set, indicating a global index.

The parent directory of MII_INDEX_FILE needs to be
created manually before generating the index.